+++ CHANGELOG.md
@@ -48,3 +48,4 @@ and `crypto_auth_verify` functions of the `NaCl` library (does not exists in Twe
 - Added SHA-256.
 - Added SHA-256 unit tests with the official testvectors.
 - Fixed some typos.
+- Added scalar_base for Ed25519Bip32 compatibility